Extract shared sanitizer helpers into PayloadSanitizationCore
`SentryPayloadSanitizer` and `AnalyticsPayloadSanitizer` both
duplicated two pieces of logic:

- `shouldDrop(key:)` lowercased the key and substring-matched against
  a sensitive-fragment list.
- `sanitizeText` ran `ObservabilityTextRedactor.redact` then capped
  the result at a per-destination max length (240 for Sentry, 80 for
  Analytics) with an ellipsis on overflow.

If the redaction or truncation policy changes, both files had to be
updated in lockstep. That's exactly the drift footgun the nightly
security check tries to catch.

Pull both helpers into `PayloadSanitizationCore` so the mechanics
live in one place while each sanitizer continues to own its own
sensitive-key list and length cap. Each call site is now a one-liner
delegating to the shared core. Public API on both sanitizers is
unchanged (`sanitizeText`, `sanitizeContext`, `sanitizeTags`,
`sanitizeAnyDictionary`, `sanitizeEventContexts`, `sanitizeProperties`,
`sanitizeDiagnosticContextForDisplay`, `redact`), and the shared
`ObservabilitySanitizerCorpus.json` fixture still drives the existing
`SentryPayloadSanitizerTests` / `AnalyticsPayloadSanitizerTests`
regression checks without modification.
